requirements.txt
Pillow
streamlit
webuiapi